requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymupdf==1.23.8
chromadb==0.4.22
sentence-transformers==2.2.2
huggingface-hub==0.16.4  # ← Specific compatible version
google-generativeai==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.24.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23